Transfer Commit
Contains some project settings & disabled network test code. Testing
tilt is not possible on the iPad simulator. This commit is to get the
latest version of the repo on my windows 8 machine to see if I can
deploy to windows phone. I have an original iPad but it cannot be
upgraded past iOS 5.1. Unfortunately Apple does not let you download
older versions of Xcode unless you are a registered developer. Perhaps
things will go smoother on the windows side.